refactor(app/registries): relocate/colocate registries related views
