+++ app/portainer/registry-management/views/create/createRegistryController.js
