+++ app/portainer/registry-management/views/edit/registryController.js
